requirements.txt
asgiref==3.2.5
asn1crypto==0.24.0
backcall==0.1.0
django-countries==6.0
django_widget_tweaks
ipython==7.13.0
ipython-genutils==0.2.0
Pillow
plotly==4.5.2
Pygments==2.6.1
requests==2.23.0
requests-unixsocket==0.1.5
retrying==1.3.3
SecretStorage==2.3.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
